refactor: split filter-panel into feature modules
