+++ lib/features/finer-filters/stat-map.ts
@@ -1,0 +1,40 @@
+export const modMap: Record<string, string> = {
+  life: "total_life",
+  cold: "total_cold_resistance",
+  fire: "total_fire_resistance",
+  light: "total_lightning_resistance",
+  chaos: "total_chaos_resistance",
+  move: "increased_movement_speed",
+  allR: "total_elemental_resistance",
+  explicit_life: "3299347043",
+  explicit_cold: "4220027924",
+  explicit_fire: "3372524247",
+  explicit_light: "1671376347",
+  explicit_chaos: "2923486259",
+  explicit_inc_phy_dmg: "1509134228",
+  explicit_add_phy_local: "1940865751",
+  explicit_add_fire_local: "709508406",
+  explicit_add_cold_local: "1037193709",
+  explicit_add_light_local: "3336890334",
+  explicit_add_chaos_local: "2223678961",
+  explicit_inc_attack_speed_local: "210067635",
+  explicit_inc_crit_chance: "2375316951",
+  explicit_global_crit_mult: "3556824919",
+  explicit_inc_spell_dmg: "2974417149",
+  explicit_inc_fire_spell_dmg: "3962278098",
+  explicit_inc_cold_spell_dmg: "3291658075",
+  explicit_inc_light_spell_dmg: "2231156303",
+  explicit_add_fire_spell_dmg: "1133016593",
+  explicit_add_cold_spell_dmg: "2469416729",
+  explicit_add_light_spell_dmg: "2831165374",
+  explicit_level_all_spells: "124131830",
+  explicit_level_all_fire_spells: "591105508",
+  explicit_level_all_cold_spells: "2254480358",
+  explicit_level_all_light_spells: "1545858329",
+  explicit_level_all_physical_spells: "4226189338",
+  explicit_level_all_chaos_spells: "2891184298",
+  explicit_inc_cast_speed: "737908626",
+  explicit_gain_extra_fire_damage: "3015669065",
+  explicit_gain_extra_cold_damage: "2505884597",
+  explicit_gain_extra_light_damage: "3278136794"
+}
